+++ sdk/metricsadvisor/Azure.AI.MetricsAdvisor/tests/Samples/Sample04_HookCrudOperations.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using Azure.AI.MetricsAdvisor.Administration;
 using Azure.AI.MetricsAdvisor.Models;
@@ -73,7 +72,7 @@ namespace Azure.AI.MetricsAdvisor.Samples
                 Console.WriteLine($" - {admin}");
             }
 
-            if (hook.HookType == HookType.Email)
+            if (hook.HookKind == NotificationHookKind.Email)
             {
                 EmailNotificationHook emailHook = hook as EmailNotificationHook;
 
@@ -83,7 +82,7 @@ namespace Azure.AI.MetricsAdvisor.Samples
                     Console.WriteLine($" - {email}");
                 }
             }
-            else if (hook.HookType == HookType.Webhook)
+            else if (hook.HookKind == NotificationHookKind.Webhook)
             {
                 WebNotificationHook webHook = hook as WebNotificationHook;
 
@@ -143,7 +142,7 @@ namespace Azure.AI.MetricsAdvisor.Samples
                 Console.WriteLine($"Hook ID: {hook.Id}");
                 Console.WriteLine($"Name: {hook.Name}");
                 Console.WriteLine($"Description: {hook.Description}");
-                Console.WriteLine($"Type: {hook.HookType}");
+                Console.WriteLine($"Type: {hook.HookKind}");
                 Console.WriteLine();
 
                 // Print at most 5 hooks.